+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- feat (guardduty): added support for malware protection, RDS, lambda and additional EKS options
+
 ### Fixed
 
 - fix(networking): fix same account cross region vpc peering route table not found